Refactor Context objects
