+++ src/renderer/objects/context/ContextSection.ts
@@ -1,0 +1,13 @@
+import ContextItem from './ContextItem'
+
+export default class ContextSection {
+  items: ContextItem[] = []
+
+  add (condition, item: ContextItem) {
+    if (condition) {
+      this.items.push(item)
+    }
+
+    return this
+  }
+}
