+++ src/renderer/objects/context/ContextTarget.ts
@@ -1,0 +1,11 @@
+export enum ContextTargetType {
+  Explorer,
+  ExplorerNode,
+  RenderedViewport,
+  ComposeViewport,
+}
+
+export default interface ContextTarget {
+  type: ContextTargetType
+  path: string
+}
