+++ src/renderer/objects/context/menus/ComposerViewportContextMenu.ts
@@ -1,11 +1,13 @@
-import { Store } from '@/store'
-import { ContextMenu, ContextItem, ContextCommand } from '@/store/modules/context'
+import { Store, State } from '@/store'
+import ContextMenu from '@/objects/context/ContextMenu'
+import ContextItem from '@/objects/context/ContextItem'
+import ContextCommand from '@/objects/context/ContextCommand'
 import { format } from '@/modules/Titles'
 
-export default function ComposerViewportContextMenu (store: Store, selection: string, replace: (string) => void) {
+export default function ComposerViewportContextMenu (store: Store<State>, selection: string, replace: (string) => void) {
   const format_interaction_titles = store.state.configuration.format_interaction_titles
 
-  return ContextMenu.define(() => [
+  return ContextMenu.define(undefined, () => [
     [
       ContextItem.menu(
         'Action',
@@ -27,7 +29,7 @@ export default function ComposerViewportContextMenu (store: Store, selection: st
         // eslint-disable-next-line @typescript-eslint/no-empty-function
         async () => {},
         ContextCommand.control().key('F'),
-      ).when(() => selection !== ''),
+      ).when(async () => selection !== ''),
     ],
     [
       ContextItem.action(
@@ -37,21 +39,21 @@ export default function ComposerViewportContextMenu (store: Store, selection: st
           replace('')
         },
         ContextCommand.control().key('X'),
-      ).when(() => selection !== ''),
+      ).when(async () => selection !== ''),
       ContextItem.action(
         'Copy',
         async () => {
           await store.dispatch('clipboard/text', selection)
         },
         ContextCommand.control().key('C'),
-      ).when(() => selection !== ''),
+      ).when(async () => selection !== ''),
       ContextItem.action(
         'Paste',
         async () => {
           await store.dispatch('clipboard/text', selection)
         },
         ContextCommand.control().key('V'),
-      ).when(() => Boolean(store.state.clipboard.content)),
+      ).when(async () => Boolean(store.state.clipboard.content)),
     ],
   ])
 }
